requirement.txt
Django==5.0.5
djangorestframework==3.14.0
django-cors-headers==4.3.1 
celery==5.3.6
redis==5.0.1
drf-yasg==1.21.7
django-environ==0.11.2
psycopg2-binary==2.9.9
